Addressable ids & change ids to uint256

+++ src/HatsOracles/IHatsOracle.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.8.13;
 
 interface IHatsOracle {
-    function checkWearerStanding(address _wearer, uint64 _hatId)
+    function checkWearerStanding(address _wearer, uint256 _hatId)
         external
         view
         returns (bool);
